requirements.txt
fastapi
uvicorn
pinecone-client
transformers
torch